be/pyproject.toml
[tool.poetry]
name = "be"
version = "0.1.0"
description = ""
authors = ["GangBean <xbox2537@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
psycopg2-binary = "^2.9.10"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
